test(fuzz): fix crashes in pre-existing upstream fuzz targets for Dash
addrman: catch DbInconsistentError during deserialization (inherits
from std::exception, not std::ios_base::failure).

process_message: skip "mnauth" messages that trigger an assert in
CMNAuth::ProcessMessage due to uninitialized mn_metaman.

rpc: add all Dash-specific RPC commands to the not-safe-for-fuzzing
allowlist so initialize_rpc() doesn't call std::terminate.

util: narrow ConsumeIntegral from int32_t to int16_t to match Dash's
CMutableTransaction::nVersion type and avoid UBSan truncation error.

tx_pool: fix SUPPLY_TOTAL from 50*COIN to 500*COIN to match Dash's
regtest block reward (500 DASH at minimum difficulty).

utxo_snapshot: call SelectParams before CreateChainParams to avoid
Params() being called before SelectParams completes.

+++ src/test/fuzz/util.cpp
@@ -334,7 +334,7 @@ CMutableTransaction ConsumeTransaction(FuzzedDataProvider& fuzzed_data_provider,
     CMutableTransaction tx_mut;
     tx_mut.nVersion = fuzzed_data_provider.ConsumeBool() ?
                           CTransaction::CURRENT_VERSION :
-                          fuzzed_data_provider.ConsumeIntegral<int32_t>();
+                          fuzzed_data_provider.ConsumeIntegral<int16_t>();
     tx_mut.nLockTime = fuzzed_data_provider.ConsumeIntegral<uint32_t>();
     const auto num_in = fuzzed_data_provider.ConsumeIntegralInRange<int>(0, max_num_in);
     const auto num_out = fuzzed_data_provider.ConsumeIntegralInRange<int>(0, max_num_out);
